Make tristate_t type generally available. JB#22475
Generic "extended boolean" type supporting als "unknown" in
addition to "true" and "false" is useful for dealing with
state data where neither true nor false is suitable for
use as initial value. There already is tristate_t type
that fits the requirement, but it is internal to the
buttonbacklight plugin.

Move tristate_t from plugin to mce core modules so that
it can be utilized anywhere within mce.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>