Track pid of topmost window. JB#41340
MCE needs to differentiate between situations where lockscreen
is shown vs. when some application is on top of the lockscreen.
This can be accomplished by tracking the pid of the application
that owns the topmost window on screen - which is -1 when
lockscreen or homescreen is on top.

Add topmost_window_pid datapipe, keep it in sync by querying
initial state from compositor on startup and then listening to
change signals.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>