Allow blanking pause when app is on top of lockscreen. JB#41340
Traditionally both user activity and blanking pause requests
have been ignored when lockscreen is active - so that the
display is powered off relatively quick unless the user
deactivates the lockscreen.

The advent of launching camera application from the lockscreen
without unlocking the device first has already required a lot
of adjusting to the rules for ignoring user activity. However
the blanking pause is still ignored, which basically means
if camera application launched from lockscreen is used for
recording a video: The display gets powered off despite
camera application trying to signal that it should be
kept on.

Allow blanking pause also when some application is on top
of the lockscreen - which is assumed to be the case when
compositor is reporting topmost window pid != -1.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>