Optional shutdown after idling. Fixes MER#1889
When doing ad-hoc testing/debugging on several devices one
often finds out that the device that is needed next needs
to be charged before it can be used.

Add option for automatically shutting down device when it
is not connected to a charger and has been idling longer
than user specified time limit.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>