Disable display state restore during bootup. JB#41395
Bootup while device is connected to a charger causes display to blank
immediately after bootup is finished. Happens on devices lipstick is
started while display is in "off" state: Display is turned on due
to charger connected notification - and is then turned back off once
the notification is removed.

Disable display state restore for ui exceptions raised during bootup.

Also, in case the exception is raised in the middle of a display state
transition, the state to be restored must be the transition target
rather than the state we are just about to leave.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>