Do not take shutdown actions on shutdown notification. JB#41302
If user attaches a charger after battery empty shutdown notification,
the device stays powered up but touch input is disabled until the
device is rebooted.

Instead of treating notifications about imminent shutdown as actual
shutdown from which there is no return, just turn the display on to
allow the user better chance of seeing the notification.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>