[mce-sensorfw] Do not retry loading of unavailable sensors. JB#41369
Now that it is possible for sensorfwd to refuse loading a sensor plugin
(due to the device not having such sensor), this can trigger MCE into
perpetual retry-in-N-seconds D-Bus IPC loop.

Differentiate between "sensor was not loaded" replies and "there were
issues while requesting sensor load" error replies - so that former
is taken to mean that the sensor is not available, and the latter as
something that might be helped with retrying a bit later.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>