Add "nop" powerkey action
If single and double press actions are exactly the same, mce does not wait
or expect to see double presses. In some cases this is desirable, in other
cases it is not.

Having a no-operation action available allows configuring single and
double press action lists so that they differ from each other while are
still logically equivalent.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>