[fingerprint] Work around fp scanner on power key issues. JB#42135
If fingerprint scanner is physically located on the power button, pressing
the button can yield both fp identification and key events - which can
result in confusion.

Delay handling of fingerprint identification a bit so that it can
be ignored if almost simultaneous power key event is also seen.

Disable fingerprint wakeup policy while power key is pressed down so that
fingerprint is either not scanned at all, or gets ignored if received.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>