Omit premature display state broadcast. JB#43703
Long time ago the intent was to broadcast display off signals at the
start of on->off transition, and display on signals at the end of
off->on transition. Over time this has changed, and one vestigial
evaluation point was left behind - which then causes switching from
on->lpm show up on dbus as on->off->lpm.

Remove unwanted broadcast checkpoint and rectify conditions for actually
doing the broadcast from other checkpoints.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>