Deny keypad input device grabbing by default. Fixes JB#43764
Previously grabbing input devices was necessary, but currently
the expected behavior is to allow compositor to see all input
events and use input policy decisions exposed on D-Bus to control
whether UI should react to key events it sees.

As grabbing can result in loss of information on UI side, takes time to
undo on display unblank and should not be required anymore, change
the built in defaults so that grabbing is not allowed.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>