Do not artificially limit PSM threshold setting
The mcetool --set-psm-threshold option uses hard-coded validity check.
While it does match the built-in default for "possible values list",
a) the list is configurable and b) it is supposed to be just a hint
to settings ui and there is no reason to limit users from using whatever
limits they want to use.

Just validate that a number that is supposed to be a percentage is
between zero and 100.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>