Do not act on initial tklock request datapipe value. JB#44605
The tklock_request_pipe does not have a "state". However using output value
tracker for the datapipe causes it to be handled as a state variable. Which
then causes the dummy datapipe default value to be handled as a tklock
change request leading to unnecessary tklock noise on mce startup.

Switch to tracking tklock_request_pipe input values.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>