Wait for p-sensor state before acting on keyboard slide open
Unblanking due to keyboard slide open needs to be subjected to proximity
sensor check. If on-demand proximity sensor power up is used, sensor state
is not available when input events are received.

Use the on-proximity action queue, so that proximity sensor gets powered
up and sensor state can be evaluated before considering unblanking.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>