Wait for p-sensor state before executing powerkey/gesture actions
The actions executed due to powerkey press / touchscreen gesture events need
to be subjected to proximity sensor check. If on-demand proximity sensor
power up is used, sensor state is not available when input events are
received.

Use the on-proximity action queue, so that proximity sensor gets powered
up and actions executed once the sensor state is known.

As conditions for acting on power key and gesture events differ, queue
predicate selection callback in front of the actual action callbacks.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>