Wait for p-sensor state before acting on D-Bus requests
When display on/dim requests are received over D-Bus, they must be
subjected to proximity sensor check. However in on-demand proximity
sensor mode this is not immediately available.

Use the on-proximity action queue, so that proximity sensor gets powered
up and display state requests are acted on once the sensor state is known.

Also refactor the method call handler functionality so that repetitive
copy-paste code can be eliminated.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>