Reject LPM mode when lockscreen is not on screen. Fixes JB#21043
When LPM is in use and gets triggered immediately after ending a call,
display gets dimmed and coloring changes but call ui is still shown on
screen (and can't be interacted with).

Reject attempts to invoke LPM state if it is expected that call or
alarm ui - which are layered above locksreen - are on screen.

When LPM state is activated, allow UI side some time to hide any
active applications. But if that does not happen, exit from LPM by
blanking screen.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>