Implement separate policy for allowing fpwakeups. Fixes JB#49634
Whether fpwakeups are allowed or not shares logic with touchscreen
gesture (e.g. doubletaps) handling. This is not entirely correct
and leads to fpwakeups getting ignored in situations where display
is on.

Implement fpwakeup specific action policy that allows event actions
to be taken when either display is off or lockscreen is active.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>